Decimal numbers to stocks

+++ Antonieta.Backend/models/ration_input_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column, Float, Integer, String, ForeignKey, DateTime
 from sqlalchemy.orm import relationship
 from database import Base
 
@@ -7,7 +7,7 @@ class RationInput(Base):
     
     id = Column(Integer, primary_key=True, index=True)
     ration_stock_id = Column(Integer, ForeignKey('ration_stock.id'), index=True)
-    amount = Column(Integer, index=True)
+    amount = Column(Float, index=True)
     date = Column(DateTime, nullable=False)
     description = Column(String, index=True)
     ration_stock = relationship("RationStock", back_populates="inputs")
